integrate upstream changes from trunk

+++ branches/ParseTree/Projects/Core/Core/Common/PrototypeConstraint.cpp
@@ -14,7 +14,7 @@ namespace ObjectiveScript {
 
 
 PrototypeConstraint::PrototypeConstraint()
-: mIndex(~0)
+: mIndex((unsigned int)~0)
 {
 }
